[refector(BookShelfViewmodel)] view관리에 대한 코드는 Fragment로 이전
